requirements.txt
apex==0.1
certifi==2020.6.20
chardet==3.0.4
cupy-cuda101==8.0.0
fastrlock==0.5
future==0.18.2
googledrivedownloader==0.4
idna==2.10
mido==1.2.9
numpy==1.19.2
Pillow==8.0.1
pretty-midi==0.2.9
requests==2.24.0
scipy==1.5.3
six==1.15.0
torch==1.6.0+cu101
tqdm==4.50.2
urllib3==1.25.11
